BOOM, Here's your physics
The physics have been implemented and bug tested to the state wheere it
doesn't fall apart when built. There might be a few fixes that needs to
be done further down the line, like adding sleeping bodies. But we live
now, and now I'm happy with this, so don't spoil it!